moving page components to page folder
