+++ webapp/client/src/components/E404/e404.js
@@ -1,0 +1,27 @@
+import React, { Component } from 'react'
+import { Link } from "react-router-dom"
+import "./e404.sass"
+
+class E404 extends Component {
+
+  componentDidMount(){
+    const game = document.getElementById("game")
+    game.style.display = "none"
+  }
+
+  render(){
+    return (
+      <div className="e404">
+        <div className="e404-title">
+          404 Unknown Page
+        </div>
+        <Link to="/" className="kts-link">
+          /home
+        </Link>
+      </div>
+    )
+  }
+
+}
+
+export default E404
