+++ webapp/client/src/components/E404/e404.sass
@@ -1,0 +1,12 @@
+.e404
+  padding-top: 100px
+  text-align: center
+  .e404-title
+    font-size: 10vw
+    color: #8D8E87
+  .kts-link
+    font-size: 30px
+    color: #ECEDE3
+    opacity: 0.5
+  .kts-link:hover
+    opacity: 1
