+++ webapp/client/src/pages/Game/game.js
@@ -7,6 +7,12 @@ class Home extends Component {
     const game = document.getElementById("game")
     game.style.display = "block"
     game.style.opacity = 1
+    console.log("process.env.NODE_ENV: ", process.env.NODE_ENV)
+    if (process.env.NODE_ENV === 'development') {
+      console.log("force resize window (development only)")
+      window.scaleLanding()
+      window.updateUI()
+    }
   }
 
   componentWillUnmount(){
